Move Config and DB functions out of DMM and DRM (#58)
* Autopep8

* Move Config / DB functions out of DMM / DRM (#58)

* Clean up imports.

* Organize app folder.
