+++ data_resource_api/api/core/versioned_resource.py
@@ -8,7 +8,7 @@ the API version number in the request header.
 from flask_restful import Resource
 from flask import request
 from data_resource_api.api.v1_0_0 import ResourceHandler as V1_0_0_ResourceHandler
-from data_resource_api.app.exception_handler import MethodNotAllowed
+from data_resource_api.app.utils.exception_handler import MethodNotAllowed
 
 
 class VersionedResourceParent(Resource):
