+++ data_resource_api/app/__init__.py
@@ -1,2 +1,2 @@
-from data_resource_api.app.data_model_manager import DataModelManager
-from data_resource_api.app.data_resource_manager import DataResourceManager
+from data_resource_api.app.data_managers.data_model_manager import DataModelManager
+from data_resource_api.app.data_managers.data_resource_manager import DataResourceManager
