+++ data_resource_api/app/utils/descriptor.py
