+++ data_resource_api/app/utils/exception_handler.py
