+++ data_resource_api/app/utils/junc_holder.py
