+++ tests/descriptor/test_descriptor.py
@@ -1,4 +1,4 @@
-from data_resource_api.app.descriptor import Descriptor
+from data_resource_api.app.utils.descriptor import Descriptor
 from tests.schemas import (
     frameworks_descriptor)
 from expects import expect, equal
@@ -16,17 +16,20 @@ class TestDescriptorClass():
     def test_get_table_name(self):
         desc = Descriptor(frameworks_descriptor)
         table_name = desc.table_name
-        expect(table_name).to(equal(frameworks_descriptor['datastore']['tablename']))
+        expect(table_name).to(
+            equal(frameworks_descriptor['datastore']['tablename']))
 
     def test_get_table_schema(self):
         desc = Descriptor(frameworks_descriptor)
         table_schema = desc.table_schema
-        expect(table_schema).to(equal(frameworks_descriptor['datastore']['schema']))
+        expect(table_schema).to(
+            equal(frameworks_descriptor['datastore']['schema']))
 
     def test_get_api_schema(self):
         desc = Descriptor(frameworks_descriptor)
         api_schema = desc.api_schema
-        expect(api_schema).to(equal(frameworks_descriptor['api']['methods'][0]))
+        expect(api_schema).to(
+            equal(frameworks_descriptor['api']['methods'][0]))
 
     def test_descriptor_with_no_file_name(self):
         desc = Descriptor(frameworks_descriptor)
